+++ framework/domain/src/main/java/is/codion/framework/domain/entity/DefaultEntityType.java
@@ -91,7 +91,7 @@ final class DefaultEntityType implements EntityType, Serializable {
 	}
 
 	@Override
-	public EntityDefinition.Builder define(List<AttributeDefinition.Builder<?, ?>> definitionBuilders) {
+	public EntityDefinition.Builder define(List<? extends AttributeDefinition.Builder<?, ?>> definitionBuilders) {
 		return new DefaultEntityDefinition.DefaultBuilder(this, requireNonNull(definitionBuilders));
 	}